+++ OpenResourceSystem/ORSResourceSuppliable.cs
@@ -6,7 +6,7 @@ using System.Text;
 namespace OpenResourceSystem {
     public interface ORSResourceSuppliable {
         void receiveFNResource(double power_supplied,String resourcename);
-        float consumeFNResource(double power_to_consume, String resourcename);
+        double consumeFNResource(double power_to_consume, String resourcename);
         float consumeFNResource(float power_to_consume, String resourcename);
         string getResourceManagerDisplayName();
         int getPowerPriority();
